+++ modules/tinymce/src/core/main/ts/commands/IndentOutdent.ts
@@ -77,7 +77,11 @@ const handle = (editor: Editor, command: string): void => {
   });
 };
 
+const indent = (editor: Editor) => handle(editor, 'indent');
+const outdent = (editor: Editor) => handle(editor, 'outdent');
+
 export {
   canOutdent,
-  handle
+  indent,
+  outdent
 };